added mail helper for feedback confirmationdefaults. updated org details.
removed commented out lines
