+++ app/views/orgs/_feedback_form.html.erb
@@ -11,11 +11,13 @@
     <div class="row">
       <div class="form-group col-xs-8">
         <%= f.label :feedback_email_subject, _('Subject'), class: "control-label" %>
-        <%= f.text_field :feedback_email_subject, class: "form-control" %>
+
+        <%= f.text_field :feedback_email_subject, class: "form-control", placeholder: _(feedback_confirmation_default_subject) % { application_name: Rails.configuration.branding[:application][:name] } %>
       </div>
     </div>
     <div class="row">
-      <div class="form-group col-xs-8" data-toggle="tooltip" data-html="true" title="<%= _('You may use the following keywords in your email: <ul><li>%{application_name} - the name of this tool</li><li>%{user_name} - the plan\'s owner</li><li>%{plan_name} - the plan\'s title</li><li>%{organisation_email} - the administrative contact email for your organisation</li></ul>') %>">
+      <% tip = _(feedback_confirmation_default_message) % { user_name: _('%{user_name}'), plan_name: _('%{plan_name}'), organisation_email: @org.contact_email } %>
+      <div class="form-group col-xs-8" data-toggle="tooltip" data-html="true" title="<%= tip %>">
         <%= f.label :feedback_email_msg, _('Message'), class: "control-label" %>
         <%= f.text_area :feedback_email_msg, class: "form-control" %>
       </div>
@@ -24,9 +26,6 @@
   <div class="row">
     <div class="form-group col-xs-8">
       <%= f.button(_('Save'), id:"save_org_submit", class: "btn btn-primary", type: "submit") %>
-      <%= f.button(_('Reset to defaults'), id:"reset-to-default-feedback-email", class: "btn btn-default", type: "button") %>
-      <div id="feedback-email-default-subject" class="hide"><%= UserMailer.feedback_confirmation_default_subject %></div>
-      <div id="feedback-email-default-message" class="hide"><%= UserMailer.feedback_confirmation_default_message %></div>
     </div>
   </div>
 <% end %>
