+++ lib/assets/javascripts/views/orgs/admin_edit.js
@@ -67,13 +67,6 @@ $(() => {
   // Initialises tinymce for any target element with class tinymce_answer
   Tinymce.init({ selector: '#org_feedback_email_msg' });
 
-  $('#reset-to-default-feedback-email').click((e) => {
-    e.preventDefault();
-    $('#org_feedback_email_subject').val($('#feedback-email-default-subject').html());
-    const editor = Tinymce.findEditorById('org_feedback_email_msg');
-    editor.setContent($('#feedback-email-default-message').text());
-  });
-
   // Convert the max number of URLs constant to text and display for user
   $('#max-nbr-urls').text(convertToText(MAX_NUMBER_ORG_URLS).toLowerCase());
